F5-TTS/f5-tts_server/requirements.txt
cached_path
f5_tts
fastapi
pydantic
pydub
python_magic
soundfile
spaces
torch==2.8.0.dev20250327+cu126
torchaudio==2.6.0.dev20250331+cu126
